Fix GCC error on Gentoo-based systems
Gentoo appears to have a funky GCC install, which doesn't seem to work
with `-fstack-protector` enabled. I have added a conditional to both
the kernel & loader Makefile to detect a Gentoo-based system (its very
much a hacky solution, I intend to make it more robust) and add
`-fno-stack-protector` to the `CFLAGS` variable.

At this moment in the time, the solution works.