Switch Makefiles to use `find` instead of `ls`
Previously, the kernel Makefile used Bash globstar and `ls **/*.c` to
find source files recursively, and the loader Makefile was not
recursive. Commit 4975091 changed the ASM_SOURCE in the kernel Makefile
to use POSIX `find` instead; this commit extends that to all the
sources, removing the dependency on Bash globstar.