Refactor kernel paging system
Major refactor of kernel paging. Uses x86 recursive page tables and
supports "huge pages" as well as e.g. PML5T in future.

Slight change to loader so that the kernel stack is not mapped to the
final PML4T entry, as that is now used for recursive page tables.